dynamic import page transition progress bar
